Refactor web cmd, remove prune commands
